+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - ASYNC100 now properly treats async for comprehensions as checkpoints.
 - ASYNC100 now supports autofixing on asyncio.
 
+## 24.4.2
+- Add ASYNC119: yield in contextmanager in async generator.
+
 ## 24.4.1
 - ASYNC91X fix internal error caused by multiple `try/except` incorrectly sharing state.